refactor: move fetchAllVaults into stellar-sdk-helpers for serverless compatibility
