+++ api/v1/vaults/index.ts
@@ -1,4 +1,4 @@
-import { fetchAllVaults } from "../../../apps/api/src/services/vaults.js";
+import { fetchAllVaults } from "@meridian/stellar-sdk-helpers";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export default async function handler(_req: any, res: any) {
