+++ packages/stellar-sdk-helpers/src/index.ts
@@ -3,3 +3,4 @@ export * from "./defilamma";
 export * from "./defindex";
 export * from "./horizon";
 export * from "./types";
+export * from "./vaults";
